Move the test to the GPU directory and mimic device=cuda in another test following the current change.
