+++ theano/gpuarray/tests/test_rng_mrg.py
@@ -12,6 +12,7 @@ from theano.sandbox.tests.test_rng_mrg import test_f16_nonzero as cpu_f16_nonzer
 from theano.tests import unittest_tools as utt
 
 from .config import mode_with_gpu as mode
+from .config import test_ctx_name
 from ..type import gpuarray_shared_constructor
 from ..rng_mrg import GPUA_mrg_uniform
 
@@ -165,4 +166,25 @@ def test_validate_input_types_gpuarray_backend():
 
 
 def test_f16_nonzero():
-    cpu_f16_nonzero(mode=mode, op_to_check=GPUA_mrg_uniform)
+    try:
+        # To have theano.shared(x
+        theano.compile.shared_constructor(gpuarray_shared_constructor)
+        cpu_f16_nonzero(mode=mode, op_to_check=GPUA_mrg_uniform)
+    finally:
+        theano.compile.shared_constructor(gpuarray_shared_constructor,
+                                          remove=True)
+
+
+def test_cpu_target_with_shared_variable():
+    srng = MRG_RandomStreams()
+    s = np.random.rand(2, 3).astype('float32')
+    x = gpuarray_shared_constructor(s, name='x')
+    y = srng.uniform(x.shape, target='cpu')
+    y.name = 'y'
+    z = (x * y).sum()
+    z.name = 'z'
+
+    fz = theano.function([], z, mode=mode)
+
+    nodes = fz.maker.fgraph.toposort()
+    assert not any([isinstance(node.op, GPUA_mrg_uniform) for node in nodes])
